seperate server/client for the analysis
